src/blog/search_indexes.py
import datetime
from haystack import indexes
from blog.models import Post, Explained

class PostIndex(indexes.SearchIndex, indexes.Indexable):
    text = indexes.CharField(document=True, use_template=True)
    author = indexes.CharField(model_attr='author')
    pub_date = indexes.DateTimeField(model_attr='published_date')

    def get_model(self):
        return Post

    def index_queryset(self, using=None):
        """Used when the entire index for model is updated."""
        return self.get_model().published.all()

class ExplainedIndex(indexes.SearchIndex, indexes.Indexable):
    text = indexes.CharField(document=True, use_template=True)
    author = indexes.CharField(model_attr='author')
    pub_date = indexes.DateTimeField(model_attr='published_date')

    def get_model(self):
        return Explained

    def index_queryset(self, using=None):
        """Used when the entire index for model is updated."""
        return self.get_model().published.all()