src/requirements.txt
Django==2.1.7
django-haystack==2.8.1
django-tinymce4-lite==1.7.4
jsmin==2.2.2
Pillow==5.4.1
psycopg2==2.8.2
pytz==2018.9
Whoosh==2.7.4
